azure-devops-advisor/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
python-jose==3.3.0
flask-restx==1.1.0
PyYAML==6.0.1
Werkzeug==2.3.7
flask-httpauth==4.8.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
